user validation created and routes to add and remove admins have been created

+++ routes/todolist.routes.js
@@ -4,6 +4,7 @@ const TodoList = require('../models/TodoList')
 const mongoose = require('mongoose')
 const deleteAllTodos = require('../controllers/deleteAllTodos')
 const validateUser = require('../controllers/validateUser')
+const validateUserIsAdmin = require('../controllers/user.controllers/validateUserIsAdmin')
 const { route } = require('./todo.routes')
 
 const router = Router()
@@ -53,7 +54,7 @@ router.post('/', async(req,res) => {
 
     try {
 
-        const newTodoList = await TodoList.create({...payload, user: userId})
+        const newTodoList = await TodoList.create({...payload, user: userId, admins: userId})
         await User.findByIdAndUpdate(userId, {$push: {todoList: newTodoList._id}})
 
         res.status(200).json(newTodoList)
@@ -107,6 +108,75 @@ router.put('/shareTodoList/:todoListId/:newUserId', async (req, res) => {
     }
 })
 
+//Remove invited users
+router.put('/removeInvitedUser/:todoListId/:InvitedUserId', async (req, res) => {
+
+    const userId = req.user.id
+    const { todoListId } = req.params
+    const { InvitedUserId } = req.params
+
+    try {
+
+        const userValidation = await TodoList.findById(todoListId)
+
+        validateUserIsAdmin(userValidation.admins, userId, 401, 'Only Admins can remove invited users')
+
+        const updatedTodoList = await TodoList.findByIdAndUpdate(todoListId, {$pull: { invitedUsers: InvitedUserId }}, { new: true })
+
+        res.status(200).json(updatedTodoList)
+        
+    } catch (error) {
+        
+        res.status(error.status || 500).json(error.message)
+    }
+})
+
+//Add new Admin to TodoList
+router.put('/addAdminTodoList/:todoListId/:newUserId', async (req, res) => {
+
+    const userId = req.user.id
+    const { todoListId } = req.params
+    const { newUserId } = req.params
+
+    try {
+
+        const validateUser = await TodoList.findById(todoListId)
+
+        validateUserIsAdmin(validateUser.admins, userId, 401, 'Only Admins can add a new Admin')
+
+        const updatedTodoList = await TodoList.findByIdAndUpdate(todoListId, {$push: { admins: newUserId }}, { new: true })
+
+        res.status(200).json(updatedTodoList)
+        
+    } catch (error) {
+        
+        res.status(error.status || 500).json(error.message)
+    }
+})
+
+//Remove Admin to TodoList
+router.put('/removeAdmin/:todoListId/:adminId', async (req, res) => {
+
+    const userId = req.user.id
+    const { todoListId } = req.params
+    const { adminId } = req.params
+
+    try {
+
+        const validateUser = await TodoList.findById(todoListId)
+
+        validateUserIsAdmin(validateUser.admins, userId, 401, 'Only Admins can remove other admins from the list')
+
+        const updatedTodoList = await TodoList.findByIdAndUpdate(todoListId, {$pull: { admins: adminId }}, { new: true })
+
+        res.status(200).json(updatedTodoList)
+        
+    } catch (error) {
+        
+        res.status(error.status || 500).json(error.message)
+    }
+})
+
 //Delete a TodoList
 router.delete('/:id', async (req, res) => {
 
@@ -117,8 +187,10 @@ router.delete('/:id', async (req, res) => {
 
         const todos = await TodoList.findById(todoListId)
 
-        validateUser(userValidation.user, userValidation.invitedUsers, userId, 401, "Cannot update another user's Todo List")
+        validateUser(todos.user, todos.invitedUsers, userId, 401, "Cannot update another user's Todo List")
 
+        validateUserIsAdmin(todos.admins, userId, 401, 'Only Admins can delete a TodoList')
+        
         deleteAllTodos(todos.todos)
 
         await TodoList.findByIdAndDelete(todoListId)
